fix(transport-problem): review fixes
- Remove unnecessary @Output
- Separate types from service
- Refactor for of to one liner
- Fix sum array helper bug
