--- apps/frontend/src/app/pages/transport-problem/table/transport-table.style.scss
@@ -1,9 +1,0 @@
-.input-table-wrapper {
-  padding: 8px;
-  display: grid;
-  grid-template-rows: auto auto;
-  grid-template-columns: auto auto;
-  gap: 16px;
-  justify-content: start;
-  overflow-x: auto;
-}
