+++ apps/frontend/src/app/pages/transport-problem/transport-problem.component.ts
@@ -1,12 +1,29 @@
 import {ChangeDetectionStrategy, Component} from '@angular/core';
 import {MatSelectChange} from '@angular/material/select';
 
-import {TPMethods, TransportProblemService} from './transport-problem.service';
+import {TransportProblemService} from './transport-problem.service';
+import {TPMethods} from './transport-table.types';
 
 @Component({
   selector: 'transport-problem-page',
   templateUrl: './transport-problem.template.html',
-  styleUrls: ['./transport-problem.style.scss'],
+  styles: [
+    `
+      mat-form-field {
+        margin: 8px;
+      }
+
+      mat-hint {
+        display: block;
+        max-width: 500px;
+        margin: 8px;
+      }
+
+      button[type='submit'] {
+        margin: 8px;
+      }
+    `,
+  ],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class TransportProblemComponent {
