--- apps/frontend/src/app/pages/transport-problem/transport-problem.style.scss
@@ -1,15 +1,0 @@
-form {
-  mat-form-field {
-    margin: 8px;
-  }
-
-  mat-hint {
-    display: block;
-    max-width: 500px;
-    margin: 8px;
-  }
-
-  button[type='submit'] {
-    margin: 8px;
-  }
-}
