+++ apps/frontend/src/app/pages/transport-problem/transport-table.types.ts
@@ -1,0 +1,17 @@
+import {Table} from '@shared/types/table.types';
+
+export type Stocks = Array<number | null>;
+export type Demands = Array<number | null>;
+export type TPMethods = 'north-west' | 'table-min' | 'vogel-korda';
+
+/** Stands for transportation problem data what contains the costs, the demand of shops and the stock of storages. */
+export interface TPData {
+  costs: Table;
+  shopDemands: Demands;
+  storageStocks: Stocks;
+}
+
+export interface Result {
+  epsilon: number;
+  table: Table;
+}
