+++ apps/frontend/src/app/shared/helpers/array.helper.ts
@@ -1,7 +1,7 @@
 /** Add numbers in an array and ignores undefined or null values. */
 export function sum(array: Array<number | null | undefined>) {
-  return array.reduce((acc, curr) => {
-    if (!acc || !curr) return curr;
-    return acc + curr;
+  return array.reduce((previous, current) => {
+    if (!previous) previous = 0;
+    return current ? previous + current : previous;
   });
 }
